requirements.txt
customtkinter
CTkMessagebox
CTkToolTip
requests
arpy
sqlitedict
tkinterweb
pillow
pyinstaller
platformdirs